test: Add Fantom Anvil config and Fantom integration test.
